fix: update scripts exposed settings and disable validation by default
This is to avoid mathml vqlidqtion conversion error when converting to epub3

+++ Common/DaisyConverterLib/Pipeline/Pipeline2/Scripts/WordToDAISY202.cs
@@ -259,24 +259,24 @@ namespace Daisy.SaveAsDAISY.Conversion.Pipeline.Pipeline2.Scripts
                                 { "No validation", "off" },
                                 { "Report validation issues", "report" },
                                 { "Abort on validation issues", "abort" },
-                            }, "abort"),
+                            }, "off"),
                         false,
                         "Whether to abort on validation issues."
                     )
                 },
-                { "validation-report", new ScriptParameter(
-                        "validation-report",
-                        "Validation reports",
-                        new PathData(
-                            PathData.InputOrOutput.output,
-                            PathData.FileOrDirectory.Directory
-                        ),
-                        false,
-                        "Output path of the validation reports",
-                        false,
-                        ParameterDirection.Output
-                    )
-                },
+                //{ "validation-report", new ScriptParameter(
+                //        "validation-report",
+                //        "Validation reports",
+                //        new PathData(
+                //            PathData.InputOrOutput.output,
+                //            PathData.FileOrDirectory.Directory
+                //        ),
+                //        false,
+                //        "Output path of the validation reports",
+                //        false,
+                //        ParameterDirection.Output
+                //    )
+                //},
                 {"language", new ScriptParameter(
                         "language",
                         "Language code",
@@ -351,10 +351,18 @@ namespace Daisy.SaveAsDAISY.Conversion.Pipeline.Pipeline2.Scripts
         public override void ExecuteScript(string inputPath)
         {
             base.ExecuteScript(inputPath);
-            if(ExtractedShapes.Count > 0) {
-                foreach (string shape in ExtractedShapes)
-                {
-                    File.Copy(shape, Path.Combine(Parameters["output"].Value.ToString(), Path.GetFileName(shape)), true);
+            if (ExtractedShapes.Count > 0) {
+                foreach (string shape in ExtractedShapes) {
+                    if (File.Exists(shape) == false) {
+                        // File has been already copied by the new word to dtbook script, continue parsing
+                        continue;
+                    }
+                    try {
+                        File.Copy(shape, Path.Combine(Parameters["output"].Value.ToString(), Path.GetFileName(shape)), true);
+                    }
+                    catch (Exception ex) {
+                        EventsHandler.onProgressMessageReceived(this, new DaisyEventArgs("Error while copying extracted shape " + shape + ": " + ex.Message));
+                    }
                 }
             }
         }
